fix(agent): keep app-server client within lint limit
